Add files via upload
The intend of this project is to use Tableau to introduce exploratory  visualization. The data I chose to explore is the data set of flight cancelation in US during 2015.
I typed the report and attached the link for the Tableau visuals I made in it. There are 3 made visualizations & all are interactive to help the viewer explore the data and dig more into it